+++ src/Aspire.Hosting.Azure/Provisioning/Internal/BaseProvisioningContextProvider.cs
@@ -3,10 +3,8 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Reflection;
 using System.Text.Json.Nodes;
 using System.Text.RegularExpressions;
-using Aspire.Hosting.Azure.Resources;
 using Aspire.Hosting.Publishing;
 using Azure;
 using Azure.Core;
@@ -47,7 +45,7 @@ internal abstract partial class BaseProvisioningContextProvider(
 
     protected readonly TaskCompletionSource _provisioningOptionsAvailable = new(TaskCreationOptions.RunContinuationsAsynchronously);
 
-    protected virtual void EnsureProvisioningOptions(JsonObject userSecrets)
+    protected void EnsureProvisioningOptions(JsonObject userSecrets)
     {
         if (!_interactionService.IsAvailable ||
             (!string.IsNullOrEmpty(_options.Location) && !string.IsNullOrEmpty(_options.SubscriptionId)))
@@ -75,86 +73,7 @@ internal abstract partial class BaseProvisioningContextProvider(
         });
     }
 
-    protected virtual async Task RetrieveAzureProvisioningOptions(JsonObject userSecrets, CancellationToken cancellationToken = default)
-    {
-        var locations = typeof(AzureLocation).GetProperties(BindingFlags.Public | BindingFlags.Static)
-                            .Where(p => p.PropertyType == typeof(AzureLocation))
-                            .Select(p => (AzureLocation)p.GetValue(null)!)
-                            .Select(location => KeyValuePair.Create(location.Name, location.DisplayName ?? location.Name))
-                            .OrderBy(kvp => kvp.Value)
-                            .ToList();
-
-        while (_options.Location == null || _options.SubscriptionId == null)
-        {
-            var messageBarResult = await _interactionService.PromptNotificationAsync(
-                 AzureProvisioningStrings.NotificationTitle,
-                 AzureProvisioningStrings.NotificationMessage,
-                 new NotificationInteractionOptions
-                 {
-                     Intent = MessageIntent.Warning,
-                     PrimaryButtonText = AzureProvisioningStrings.NotificationPrimaryButtonText
-                 },
-                 cancellationToken)
-                 .ConfigureAwait(false);
-
-            if (messageBarResult.Canceled)
-            {
-                // User canceled the prompt, so we exit the loop
-                _provisioningOptionsAvailable.SetException(new MissingConfigurationException("Azure provisioning options were not provided."));
-                return;
-            }
-
-            if (messageBarResult.Data)
-            {
-                var result = await _interactionService.PromptInputsAsync(
-                    AzureProvisioningStrings.InputsTitle,
-                    AzureProvisioningStrings.InputsMessage,
-                    [
-                        new InteractionInput { Name = LocationName, InputType = InputType.Choice, Label = AzureProvisioningStrings.LocationLabel, Placeholder = AzureProvisioningStrings.LocationPlaceholder, Required = true, Options = [..locations] },
-                        new InteractionInput { Name = SubscriptionIdName, InputType = InputType.SecretText, Label = AzureProvisioningStrings.SubscriptionIdLabel, Placeholder = AzureProvisioningStrings.SubscriptionIdPlaceholder, Required = true },
-                        new InteractionInput { Name = ResourceGroupName, InputType = InputType.Text, Label = AzureProvisioningStrings.ResourceGroupLabel, Value = GetDefaultResourceGroupName() },
-                    ],
-                    new InputsDialogInteractionOptions
-                    {
-                        EnableMessageMarkdown = true,
-                        ValidationCallback = static (validationContext) =>
-                        {
-                            var subscriptionInput = validationContext.Inputs[SubscriptionIdName];
-                            if (!Guid.TryParse(subscriptionInput.Value, out var _))
-                            {
-                                validationContext.AddValidationError(subscriptionInput, AzureProvisioningStrings.ValidationSubscriptionIdInvalid);
-                            }
-
-                            var resourceGroupInput = validationContext.Inputs[ResourceGroupName];
-                            if (!IsValidResourceGroupName(resourceGroupInput.Value))
-                            {
-                                validationContext.AddValidationError(resourceGroupInput, AzureProvisioningStrings.ValidationResourceGroupNameInvalid);
-                            }
-
-                            return Task.CompletedTask;
-                        }
-                    },
-                    cancellationToken).ConfigureAwait(false);
-
-                if (!result.Canceled)
-                {
-                    _options.Location = result.Data[LocationName].Value;
-                    _options.SubscriptionId = result.Data[SubscriptionIdName].Value;
-                    _options.ResourceGroup = result.Data[ResourceGroupName].Value;
-                    _options.AllowResourceGroupCreation = true; // Allow the creation of the resource group if it does not exist.
-
-                    var azureSection = userSecrets.Prop("Azure");
-
-                    // Persist the parameter value to user secrets so they can be reused in the future
-                    azureSection["Location"] = _options.Location;
-                    azureSection["SubscriptionId"] = _options.SubscriptionId;
-                    azureSection["ResourceGroup"] = _options.ResourceGroup;
-
-                    _provisioningOptionsAvailable.SetResult();
-                }
-            }
-        }
-    }
+    protected abstract Task RetrieveAzureProvisioningOptions(JsonObject userSecrets, CancellationToken cancellationToken = default);
 
     [GeneratedRegex(@"^[a-zA-Z0-9_\-\.\(\)]+$")]
     private static partial Regex ResourceGroupValidCharacters();
